Move active fps calca to finality mod
