+++ contracts/btc-staking/src/staking.rs
@@ -2,17 +2,16 @@ use bitcoin::absolute::LockTime;
 use bitcoin::consensus::deserialize;
 use bitcoin::hashes::Hash;
 use bitcoin::{Transaction, Txid};
-use cosmwasm_std::{DepsMut, Env, Event, MessageInfo, Order, Response, Storage};
+use cosmwasm_std::{DepsMut, Env, Event, MessageInfo, Response, Storage};
 use hex::ToHex;
 
 use std::str::FromStr;
 
 use crate::error::ContractError;
-use crate::msg::FinalityProviderInfo;
 use crate::state::config::{ADMIN, CONFIG, PARAMS};
 use crate::state::staking::{
     fps, BtcDelegation, FinalityProviderState, ACTIVATED_HEIGHT, DELEGATIONS, DELEGATION_FPS, FPS,
-    FP_DELEGATIONS, FP_SET, TOTAL_POWER,
+    FP_DELEGATIONS,
 };
 use crate::validation::{
     verify_active_delegation, verify_new_fp, verify_slashed_delegation, verify_undelegation,
@@ -332,47 +331,6 @@ fn btc_undelegate(
     //  - How to notify them? Emit event?
 
     // TODO? Record event that the BTC delegation becomes unbonded at this height
-
-    Ok(())
-}
-
-/// `compute_active_finality_providers` sorts all finality providers, counts the total voting
-/// power of top finality providers, and records them in the contract state
-pub fn compute_active_finality_providers(
-    storage: &mut dyn Storage,
-    env: Env,
-    max_active_fps: usize,
-) -> Result<(), ContractError> {
-    // Sort finality providers by power
-    let (finality_providers, running_total): (_, Vec<_>) = fps()
-        .idx
-        .power
-        .range(storage, None, None, Order::Descending)
-        .take(max_active_fps)
-        .scan(0u64, |acc, item| {
-            let (pk_hex, fp_state) = item.ok()?; // Error ends the iteration
-
-            let fp_info = FinalityProviderInfo {
-                btc_pk_hex: pk_hex,
-                power: fp_state.power,
-            };
-            *acc += fp_state.power;
-            Some((fp_info, *acc))
-        })
-        .filter(|(fp, _)| {
-            // Filter out FPs with no voting power
-            fp.power > 0
-        })
-        .unzip();
-
-    // TODO: Online FPs verification
-    // TODO: Filter out slashed / offline / jailed FPs
-    // Save the new set of active finality providers
-    // TODO: Purge old (height - finality depth) FP_SET entries to avoid bloating the storage
-    FP_SET.save(storage, env.block.height, &finality_providers)?;
-    // Save the total voting power of the top n finality providers
-    let total_power = running_total.last().copied().unwrap_or_default();
-    TOTAL_POWER.save(storage, &total_power)?;
 
     Ok(())
 }
